fixed code

+++ routes/api/product-routes.js
@@ -16,7 +16,6 @@ const includeCategory = {
 
 const includeTags = {
     model: Tag,
-    as: 'tagged_products',
     attributes: [
         'id',
         'tag_name',
@@ -50,7 +49,7 @@ router.post('/', (req, res) => {
     Product.create(req.body)
         .then((product) => {
             // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-            if (req.body.tagIds.length) {
+            if (req.body.tagIds?.length) {
                 const productTagIdArr = req.body.tagIds.map((tag_id) => {
                     return {
                         product_id: product.id,
@@ -83,26 +82,32 @@ router.put('/:id', (req, res) => {
         })
         .then((productTags) => {
             // get list of current tag_ids
-            const productTagIds = productTags.map(({ tag_id }) => tag_id);
-            // create filtered list of new tag_ids
-            const newProductTags = req.body.tagIds
-                .filter((tag_id) => !productTagIds.includes(tag_id))
-                .map((tag_id) => {
-                    return {
-                        product_id: req.params.id,
-                        tag_id,
-                    };
-                });
-            // figure out which ones to remove
-            const productTagsToRemove = productTags
-                .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-                .map(({ id }) => id);
+        if(productTags?.length>0){
 
-            // run both actions
-            return Promise.all([
-                ProductTag.destroy({ where: { id: productTagsToRemove } }),
-                ProductTag.bulkCreate(newProductTags),
-            ]);
+                const productTagIds = productTags.map(({ tag_id }) => tag_id);
+                // create filtered list of new tag_ids
+                if(req.body.tagIds?.length>0){
+                    const newProductTags = req.body.tagIds
+                        .filter((tag_id) => !productTagIds.includes(tag_id))
+                        .map((tag_id) => {
+                            return {
+                                product_id: req.params.id,
+                                tag_id,
+                            };
+                        });
+                    // figure out which ones to remove
+                    const productTagsToRemove = productTags
+                        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+                        .map(({ id }) => id);
+
+                    // run both actions
+                    return Promise.all([
+                        ProductTag.destroy({ where: { id: productTagsToRemove } }),
+                        ProductTag.bulkCreate(newProductTags),
+                    ]);
+                }
+            }
+
         })
         .then((updatedProductTags) => res.json(updatedProductTags))
         .catch((err) => {
